+++ app/components/header/items_group.go
@@ -1,0 +1,31 @@
+package header
+
+import "strings"
+
+type ItemsGroup struct {
+	Items Items
+	delim string
+}
+
+func NewItemsGroup(items ...Item) ItemsGroup {
+	return ItemsGroup{
+		Items: items,
+		delim: "\\quad{}",
+	}
+}
+
+func (i ItemsGroup) Display() string {
+	list := make([]string, 0, len(i.Items))
+
+	for _, item := range i.Items {
+		list = append(list, item.Display())
+	}
+
+	return strings.Join(list, i.delim)
+}
+
+func (i ItemsGroup) Delim(delim string) ItemsGroup {
+	i.delim = delim
+
+	return i
+}
